Add thumbnail version to offer images
This change also includes a change to how different 'views' of offers
are held in memory. Different 'views' mean different structs for how
the offers should be encoded into JSON, how they should be stored in
the DB and what kind of data we expect in case of a POST/PUT request
for an offer.