Add model object for offer group post
This object will be used to group offers for FB posts. This way we
don't spam the restaurants FB page stream with a bunch of offers every
day. Instead we group all the days offers and gather them into a single
FB post.

Also, the `dateWithoutTime` is unexported (lower-cased first letter)
because this way we can make sure that no random strings are inserted
through that field.