Refactor posting to FB into a separate function
The idea is to have a `updateFBPost` sort of method that can be called
whenever an offer is either created or changed. This can then be
changed to create a group post for all of the day's offers and can
later also be used for when a group post's message template is changed
etc.